+++ src/backend/OrderBookService.ApiTests/TestInfrastructure/OrderBookTestFixture.cs
@@ -1,5 +1,6 @@
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.Extensions.Hosting;
+using OrderBookService;
 
 
 public class OrderBookTestFixture : WebApplicationFactory<Program>
