+++ src/backend/OrderBookService/DependencyInjection/ServiceCollectionExtensions.cs
@@ -1,4 +1,4 @@
-using OrderBookService.Services.ApplicationServices;
+using OrderBookService.Domain.Services.OrderBook;
 
 namespace OrderBookService.DependencyInjection;
 
@@ -6,7 +6,7 @@ public static class ServiceCollectionExtensions
 {
 	public static IServiceCollection AddOrderBookServices(this IServiceCollection services)
 	{
-		services.AddTransient<IOrderbookService, Services.ApplicationServices.OrderBookService>();
+		services.AddTransient<IOrderBookService, Domain.Services.OrderBook.OrderBookService>();
 
 		return services;
 	}
