+++ src/backend/OrderBookService/Domain/Models/AssetClasses/CoinPair.cs
@@ -1,4 +1,4 @@
-namespace OrderBookService.Models;
+namespace OrderBookService.Domain.Models.AssetClasses;
 
 internal record CoinPair: AssetClassBase
 {
