+++ src/backend/OrderBookService/Domain/Models/Orders/OrderAction.cs
@@ -1,0 +1,7 @@
+namespace OrderBookService.Domain.Models.Orders;
+
+internal enum OrderAction
+{
+	Buy,
+	Sell
+}
