--- src/backend/OrderBookService/Domain/Orders/OrderAction.cs
@@ -1,7 +1,0 @@
-namespace OrderBookService.Models;
-
-internal enum OrderAction
-{
-	Buy,
-	Sell
-}
