+++ src/backend/OrderBookService/OrderBookService.csproj
@@ -26,4 +26,8 @@
         <PackageReference Include="JetBrains.Annotations" Version="2022.1.0" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Domain\Models" />
+    </ItemGroup>
+
 </Project>
